Update and rename test.rs to tests.rs
Removes the #[cfg(test)] from the file and removed extra indentation. Also renamed it to tests.rs as asked.
